wallet/http: create open endpoint
This commit adds an endpoint POST /wallet/:id/open.
It accepts options similar to POST /wallet/:id/create.

It is more flexible than the RPC endpoint sendopen
because it accepts `broadcast` and `sign` parameters.
`broadcast` and `sign` both default to true. This allows
for the wallet to create an open transaction and choose
whether it wants to broadcast it or not sign it. It will
fail if `broadcast: true` and `sign: false`. Turning
off signing is useful for watch only wallets.