wallet/http: test account parameter, open endpoint
This PR adds tests for the account parameter when
creating OPEN transactions via the REST API. It is
important for applications built on top of hsd
to be able to specify which wallet account they are
spending from when creating transactions. This work
lays the foundation for additional endpoints that
allow for a flexible REST API regarding the auctions.