wallet: account polymorphic type check
This commit allows the `acct` parameter to the wallet
method `makeOpen` accept either an integer or a
string. The code originally asserted that it was an
integer, but this was changed because the functions
down the stack that accept `acct` are polymorphic and
will convert a human readable "account name" into an
accountIndex, which is an integer. This change allows
for a specific human readable account name to be sent
from the client which will be used to restrict coin
selection along with selecting the correct change address.